Modify Delete comment
대댓이 없으면 바로 삭제, 있으면 isDeleted를 true로 바꿈

+++ src/main/java/com/lucete/template/info/service/CommentService.java
@@ -56,7 +56,13 @@ public class CommentService {
 
     public void deleteComment(Long id) {
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Comment", "id", id));
-        commentRepository.delete(comment);
+        List<Comment> childComments= commentRepository.findByParent(id);
+        if(childComments.isEmpty())
+            commentRepository.delete(comment);
+        else {
+            comment.setIsDeleted(true);
+            commentRepository.save(comment);
+        }
     }
 
     public List<CommentDTO> getAllComments() {
